reorganize by subject
