+++ src/templates/dates/index.js
@@ -1,6 +1,6 @@
 const misc = require('./misc');
 const parsers = require('./parsers');
-const pipeSplit = require('../../_parsers/pipeSplit');
+const pipeSplit = require('../_parsers/pipeSplit');
 const timeSince = require('./_timeSince');
 const date = parsers.date;
 const natural_date = parsers.natural_date;
