+++ src/templates/formatting/format.js
@@ -104,5 +104,4 @@ inline.forEach((k) => {
   };
 });
 
-
 module.exports = templates;
