+++ src/templates/formatting/index.js
@@ -1,0 +1,5 @@
+module.exports = Object.assign({},
+  require('./format'),
+  require('./lists'),
+  require('./misc'),
+);
