+++ src/templates/formatting/misc.js
@@ -87,21 +87,7 @@ const inline = {
     }
     return '';
   },
-  //https://en.wikipedia.org/wiki/Template:Marriage
-  //this one creates a template, and an inline response
-  marriage: (tmpl, r) => {
-    let data = pipeSplit(tmpl, ['name', 'from', 'to', 'end']);
-    r.templates.push(data);
-    let str = `${data.name || ''}`;
-    if (data.from) {
-      if (data.to) {
-        str += ` (m. ${data.from}-${data.to})`;
-      } else {
-        str += ` (m. ${data.from})`;
-      }
-    }
-    return str;
-  },
+
   //https://en.wikipedia.org/wiki/Template:Lbs
   lbs: (tmpl) => {
     let obj = pipeSplit(tmpl, ['text']);
@@ -120,7 +106,5 @@ const inline = {
 
 inline['str left'] = inline.trunc;
 inline['str crop'] = inline.trunc;
-
-
 
 module.exports = inline;
