+++ src/templates/language/index.js
@@ -1,0 +1,5 @@
+module.exports = Object.assign({},
+  require('./languages'),
+  require('./pronounce'),
+  require('./wiktionary'),
+);
