+++ src/templates/language/wiktionary.js
