+++ src/templates/money/index.js
