+++ src/templates/sports/index.js
@@ -1,0 +1,2 @@
+let sports = {};
+module.exports = sports;
